@@ -17,18 +17,18 @@
     "type": "symfony-bundle",
     "minimum-stability": "dev",
     "require": {
-        "php": ">=7.4",
+        "php": ">=8.0",
         "symfony/framework-bundle": "5.4.*",
-        "roadiz/core-bundle": "~2.0.36 || dev-develop",
-        "roadiz/rozier": "~2.0.12 || dev-develop",
-        "roadiz/compat-bundle": "~2.0.0 || dev-develop",
-        "roadiz/openid": "~2.0.0"
+        "roadiz/core-bundle": "^2.2",
+        "roadiz/rozier": "^2.2",
+        "roadiz/compat-bundle": "^2.2",
+        "roadiz/openid": "^2.2"
     },
     "require-dev": {
         "php-coveralls/php-coveralls": "^2.4",
         "phpstan/phpstan": "^1.5.3",
         "phpstan/phpstan-symfony": "^1.1.8",
-        "phpstan/phpstan-doctrine": "^1.3.1",
+        "phpstan/phpstan-doctrine": "^1.3",
         "squizlabs/php_codesniffer": "^3.5"
     },
     "config": {
@@ -39,7 +39,8 @@
         "sort-packages": true,
         "allow-plugins": {
             "symfony/flex": false,
-            "symfony/runtime": false
+            "symfony/runtime": false,
+            "php-http/discovery": false
         }
     },
     "autoload": {
@@ -49,5 +50,11 @@
         "files": [
             "deprecated.php"
         ]
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-main": "2.0.x-dev",
+            "dev-develop": "2.2.x-dev"
+        }
     }
 }
