@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace RZ\Roadiz\RozierBundle\Controller\Document;
 
 use Doctrine\Persistence\ManagerRegistry;
-use RZ\Roadiz\Core\Events\DocumentUpdatedEvent;
 use RZ\Roadiz\CoreBundle\Entity\Document;
+use RZ\Roadiz\Documents\Events\DocumentUpdatedEvent;
 use RZ\Roadiz\RozierBundle\Form\DocumentLimitationsType;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\Request;
